merges changes in trunk 420:HEAD
git-svn-id: https://hrpsys-base.googlecode.com/svn/tags/3.1.3@422 a991ac11-fb38-5095-8c12-a1ddb0715245

+++ rtc/RobotHardware/robot.cpp
@@ -545,7 +545,7 @@ size_t robot::lengthOfExtraServoState(int id)
     return length_of_extra_servo_state(id);
 }
 
-void robot::readExtraServoState(int id, long *state)
+void robot::readExtraServoState(int id, int *state)
 {
     read_extra_servo_state(id, state);
 }
